refactor: move installMultiple to install/
