+++ tsconfig.json
@@ -54,11 +54,11 @@
     "src/index.ts",
     "src/install/flattenDependencies.ts",
     "src/install/index.ts",
+    "src/install/installMultiple.ts",
     "src/install/isAvailable.ts",
     "src/install/linkBins.ts",
     "src/install/linkPeers.ts",
     "src/install/postInstall.ts",
-    "src/installMultiple.ts",
     "src/logger/index.ts",
     "src/logger/pretty.ts",
     "src/logger/simple.ts",
